Changing Settings
I created a settings folder and I added three files, base, local and production. These files contain all the configuration needed to run
a django application, but this structure was used in order to divide the settings in a logical order for developing and production.

I had to change wsgi, asgi and manage and let them know the new settings file, in this case, is local for production