First views
I created a folder called api, this contains the serializers.py, urls.py and views.py used for this users application.
The serializer is based on the User model and this will convert querysets and model fields to Python dictionaries in
order to send JSON to the browser as responses.

Views:
There are two main kind of views: class based views and function based views. I used two functions, so DRF documentation
tells me that I had to use the api_view decorator and pass in a list with the corresponding HTTP methods that this view
is going to handle. Then, I just used the request.method to divide de code and do whatever the HTTP request method
is supposed to do.