Password functionality
In order to encrypt the passwords when the API is trying to create or update a database record, I had to use the create and update methods.
The save method knows when the serializer has to call update or create depending on the object attributes (whether or not an instance
argument was placed in the serializer construction). If the serializers try to add a new record, calls the create method and this
method receives the validated data. Using the "**" Python notation, it can create a new User object passing the dictionary key value
pairs as arguments and then, using the set_password method, the model overwrites the current password for another one that is encrypted.

The update method is similar. It receives the previous model instance and the validated_data. Using the super update method, the instance
fields can be updated automatically, then, the serializer can set the password the same way as it did in the create method. Both methods
return a user object.