Base and Products applications
The base app is used for settings and utils that others applications are going to use, for example, this app contains the BaseModel
that is the super class for all the ptoject models. This base model inherits from models.Models and has common model fields that
the other models take for free without repeating that fields, making the code cleaner. The repeated fields are the id, state and some
dates for creation, update and delete. In order to make this BaseModel as the super class, I needed to add the Meta class (used to
set configuration, anything that is not a field) and specify abstract as True. That way, that model is an abstract class, so Django
knows is just a parent class with common information and it doesn't create a database table for that super model.

The products app models inherit from that BaseModel and they specify their own fields that are added to the base fields. Each model has
a historical attrbute in order to keep some history information about the database operations. The method _user_history is defined
with a property decorator to get changed_by as an attribute and the model also has a setter decorator for that propery, in order to
set which user changed the record with just an assignment operation.