First API List View
In general_views I created some api views to handle generic logic for secondary views and models. This file contains three list views for
the MeasuerUnit, CategoryProduct and Indicator models. That are generic views provided by DRF, so the view can be written as soon as posible
and handle a common used pattern as get a list of database fields and send it to the user.

Each class inherit from ListAPIView, a generic view that provides a read only endpoint to get a collection of model instances. This view provides
a list method that lists a queryset. This query has to be specified in the get_queryset method with a QuerySet object being returned. If the
queryset is successfully completed, the view sends a serialized representation of the queryset as response. The last specification
that ListView needs is serializer_class, in order to let it know which serializer do I want to use.