BaseListAPIView
Since that generic_views.py has a lot of classes that repeat code, I created a base class that inherits from ListAPIView in api.py. This class
defines the get_queryset method and access the serializer model. In order to do this, I can use the get_serializer method that creates an
instance of the serializer class. Within that instance, I access the Meta class and the model defined in there. Finally, I use the context
manager (model.objects) filtering the data by the state. Each class that inherits from this base class doesn't need to create the get_queryset
method. The requisite is define serializer_class as the attribute and that view will retrieve all the list items that suit the queryset condition