CreateAPIView and Serializer Representation
I made a ProductCreateAPIView that inherits from CreateAPIView in order to set and endpoint for create only functionality.
This class only needs to declare "serializer_class" as attribute and it will receive a POST request, pass the incoming JSON
data to the serializer, check the model and serialzer validations and finally returns a response "HTTP 201" and the
serialized representation of the object if the object was successfully created, otherwise, it returns "HTTP_400" with
error details.

The ProductSerializer has a method called to_representaion that takes each product model instance and returns the desired
JSON representation. This representation will contain the id, name, description, image and some foreign records, such as
category and measure_unit. I decided to access the description field of that records, using that representation for them.