Update, Read and Delete Functionality
Django Rest Framework provides three generic views for handling common operations as update, read and delete in a database.
For update, I used UpdateAPIView with its serializer class as ProductSerializer. It also needs a get_queryset method
to tell Django the queryset where it performs the id (passed through the URL) filtering. This view also has the "patch"
method for getting the record information and the "put" method for handling extra logic in this generic view.

The other ones are practically the same, Delete and Retrieve API views need a serializer_class as attribute and the get_queryset
method to filter that query by id. The "delete" method in ProductDeleteAPIView was overwritten because I didn't want to delete
the record from the database, just put its state as False and save that. If the object exists it returns a success message, otherwise
it returns 404